# src/main/resources/openapi/identity-api.yaml
openapi: 3.0.3

info:
  title: Identity Service API
  description: API-first identity service with versioning and contract-driven design
  version: 2.0.0

servers:
  - url: http://localhost:8080/api

tags:
  - name: Users
    description: User management

paths:

  ########################
  # V1
  ########################

  /v1/users:
    post:
      tags: [Users]
      summary: Create a new user (v1)
      operationId: createUserV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestV1'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID (v1)
      operationId: getUserByIdV1
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user (soft delete) (v1)
      operationId: deleteUserV1
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  ########################
  # V2 (evoluzione)
  ########################

  /v2/users:
    post:
      tags: [Users]
      summary: Create a new user (v2)
      operationId: createUserV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestV2'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseV2'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /v2/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID (v2)
      operationId: getUserByIdV2
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseV2'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Users]
      summary: Update user partially (v2)
      operationId: updateUserV2
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequestV2'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseV2'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  ########################
  # COMPONENTS
  ########################

components:

  ########################
  # PARAMETERS
  ########################

  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  ########################
  # SCHEMAS
  ########################

  schemas:

    ###########
    # V1 DTOs
    ###########

    CreateUserRequestV1:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    UserResponseV1:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time

    ###########
    # V2 DTOs (evoluzione)
    ###########

    CreateUserRequestV2:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1

    UpdateUserRequestV2:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'

    UserResponseV2:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ###########
    # SHARED
    ###########

    UserStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - SUSPENDED

    ###########
    # RFC7807
    ###########

    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

  ########################
  # RESPONSES
  ########################

  responses:

    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    Conflict:
      description: Conflict (e.g. email already exists)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'